+++ process/drivers/gitlab_driver.rs
@@ -47,7 +47,10 @@ impl CiDriver for GitlabDriver {
 
     fn generate_tags(opts: &GenerateTagsOpts) -> miette::Result<Vec<String>> {
         const MR_EVENT: &str = "merge_request_event";
-        let os_version = Driver::get_os_version(opts.oci_ref)?;
+        let os_version = Driver::get_os_version()
+            .oci_ref(opts.oci_ref)
+            .platform(opts.platform)
+            .call()?;
         let timestamp = blue_build_utils::get_tag_timestamp();
         let short_sha =
             get_env_var(CI_COMMIT_SHORT_SHA).inspect(|v| trace!("{CI_COMMIT_SHORT_SHA}={v}"))?;
@@ -299,6 +302,7 @@ mod test {
             &GenerateTagsOpts::builder()
                 .oci_ref(&oci_ref)
                 .maybe_alt_tags(alt_tags)
+                .platform(crate::drivers::types::Platform::LinuxAmd64)
                 .build(),
         )
         .unwrap();