+++ process/drivers/local_driver.rs
@@ -23,7 +23,10 @@ impl CiDriver for LocalDriver {
 
     fn generate_tags(opts: &GenerateTagsOpts) -> miette::Result<Vec<String>> {
         trace!("LocalDriver::generate_tags({opts:?})");
-        let os_version = Driver::get_os_version(opts.oci_ref)?;
+        let os_version = Driver::get_os_version()
+            .oci_ref(opts.oci_ref)
+            .platform(opts.platform)
+            .call()?;
         let timestamp = blue_build_utils::get_tag_timestamp();
         let short_sha = commit_sha();