+++ process/drivers/opts/inspect.rs
@@ -2,11 +2,17 @@ use std::borrow::Cow;
 
 use bon::Builder;
 
+use crate::drivers::types::Platform;
+
 #[derive(Debug, Clone, Builder)]
+#[builder(derive(Clone))]
 pub struct GetMetadataOpts<'scope> {
     #[builder(into)]
     pub image: Cow<'scope, str>,
 
     #[builder(into)]
     pub tag: Option<Cow<'scope, str>>,
+
+    #[builder(default)]
+    pub platform: Platform,
 }