+++ process/drivers/types.rs
@@ -170,12 +170,48 @@ impl DetermineDriver<CiDriverType> for Option<CiDriverType> {
     }
 }
 
-#[derive(Deserialize, Debug, Clone)]
-pub struct ImageMetadata {
-    #[serde(alias = "Labels")]
-    pub labels: HashMap<String, Value>,
+#[derive(Debug, Default, Clone, Copy, ValueEnum, PartialEq, Eq, Hash)]
+pub enum Platform {
+    #[default]
+    #[value(name = "native")]
+    Native,
+    #[value(name = "linux/amd64")]
+    LinuxAmd64,
 
-    #[serde(alias = "Digest")]
+    #[value(name = "linux/arm64")]
+    LinuxArm64,
+}
+
+impl Platform {
+    /// The architecture of the platform.
+    #[must_use]
+    pub const fn arch(&self) -> &str {
+        match *self {
+            Self::Native => "native",
+            Self::LinuxAmd64 => "amd64",
+            Self::LinuxArm64 => "arm64",
+        }
+    }
+}
+
+impl std::fmt::Display for Platform {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match *self {
+                Self::Native => "native",
+                Self::LinuxAmd64 => "linux/amd64",
+                Self::LinuxArm64 => "linux/arm64",
+            }
+        )
+    }
+}
+
+#[derive(Deserialize, Debug, Clone)]
+#[serde(rename_all = "PascalCase")]
+pub struct ImageMetadata {
+    pub labels: HashMap<String, Value>,
     pub digest: String,
 }